refactor(killswitch): split firewall.rs into per-backend submodules
Decompose the ~86KB firewall.rs coordinator into focused submodules to

improve readability of security-critical code without changing behavior.

New layout under src/killswitch/firewall/:

  - mod.rs       coordinator: KillSwitch struct, public API, enable/disable,

                 update, backend selection, VPN-IP detection, Drop fail-closed,

                 and inline tests.

  - error.rs     KillSwitchError enum.

  - iptables.rs  iptables/iptables-legacy command execution: run_single_script,

                 legacy_variant, check_sudo_access, check_iptables_legacy_access,

                 verify_rules_exist.

  - ip6tables.rs IPv6 helpers: build_ipv6_script and IPV6_OUTPUT_RULES patterns.

  - nftables.rs  nft backend: nft_is_available, verify_nft_rules_exist, enable_nft,

                 disable_nft, cleanup_nft_table, run_nft, build_nft_ruleset, and

                 the nft ruleset test mod.

  - chains.rs    robust_iptables_cleanup duplicate-rule sweeper.

  - builder.rs   build_complete_script, build_rules_preview, build_dns_rules,

                 build_doh_blocking_rules.

All items keep their original visibility (pub / pub(super) / private).

Public API path crate::killswitch::firewall::{KillSwitch, KillSwitchError,

KillSwitchStatus} is preserved via pub use error::KillSwitchError.

Every SPDX header is preserved. Every SHROUD-VULN-* comment is preserved.

tests/regression.rs include_str! paths updated to the new locations.

Zero behavior change. cargo build / clippy -D warnings / test --all-features

all pass: 259+778+35+15+12 (baseline).
