+++ src/killswitch/firewall/mod.rs
@@ -29,15 +29,24 @@
 //! - `tunnel`: Allow IPv6 only via VPN tunnel interfaces
 //! - `off`: No special IPv6 handling (legacy)
 
+// Submodules — see module-level doc comments in each for ownership.
+mod builder;
+mod chains;
+mod error;
+mod ip6tables;
+mod iptables;
+mod nftables;
+
 use std::net::IpAddr;
 use std::process::Stdio;
 use std::time::Instant;
-use thiserror::Error;
 use tokio::process::Command;
 use tracing::{debug, info, warn};
 
 use crate::config::{DnsMode, Ipv6Mode};
-use crate::killswitch::paths::{ip6tables, iptables, nft};
+use crate::killswitch::paths::{iptables, nft};
+
+pub use error::KillSwitchError;
 
 /// Name of the iptables chain for the kill switch
 const CHAIN_NAME: &str = "SHROUD_KILLSWITCH";
@@ -49,45 +58,10 @@ const NFT_TABLE: &str = "shroud_killswitch";
 /// Prevents race conditions from rapid enable/disable
 const TOGGLE_COOLDOWN_MS: u64 = 500;
 
-/// Re-export for local use — single source of truth is `rules::DOH_PROVIDERS`.
-use crate::killswitch::rules::DOH_PROVIDERS as DOH_PROVIDER_IPS;
-
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 enum FirewallBackend {
     Iptables,
     Nftables,
-}
-
-/// Errors that can occur during kill switch operations.
-#[derive(Error, Debug)]
-#[allow(clippy::enum_variant_names)]
-pub enum KillSwitchError {
-    /// iptables is not installed or not in PATH
-    #[error("iptables is not available. Install with: sudo apt install iptables")]
-    #[allow(dead_code)]
-    NotFound,
-
-    /// Permission denied - need elevated privileges
-    #[error(
-        "Permission denied. Kill switch requires sudo access. Run: ./setup.sh --install-sudoers"
-    )]
-    Permission,
-
-    /// Failed to spawn iptables/sudo process
-    #[error("Failed to spawn iptables process: {0}")]
-    Spawn(#[source] std::io::Error),
-
-    /// iptables command returned error
-    #[error("iptables command failed: {0}")]
-    Command(String),
-
-    /// Failed to write to process stdin
-    #[error("Failed to write to process: {0}")]
-    Write(#[source] std::io::Error),
-
-    /// Failed waiting for iptables process
-    #[error("Failed waiting for iptables process: {0}")]
-    Wait(#[source] std::io::Error),
 }
 
 /// Kill switch status
@@ -424,441 +398,6 @@ impl KillSwitch {
         Ok(())
     }
 
-    fn build_complete_script(&self, vpn_ips: &[IpAddr]) -> String {
-        let mut s = String::new();
-
-        // Note: Cleanup is handled by robust_iptables_cleanup() before this is called
-        // We just need to create the chain and add rules
-
-        // Create chain (no jump rule yet — chain must be fully populated first
-        // to avoid a window where traffic hits an incomplete chain)
-        s.push_str(&format!("{} -N SHROUD_KILLSWITCH\n", iptables()));
-
-        // Rules
-        s.push_str(&format!(
-            "{} -A SHROUD_KILLSWITCH -o lo -j ACCEPT\n",
-            iptables()
-        ));
-        s.push_str(&format!(
-            "{} -A SHROUD_KILLSWITCH -m conntrack --ctstate ESTABLISHED,RELATED -j ACCEPT\n",
-            iptables()
-        ));
-        // DNS rules must come before VPN interface allow rules
-        s.push_str(&self.build_dns_rules());
-
-        s.push_str(&format!(
-            "{} -A SHROUD_KILLSWITCH -o tun+ -j ACCEPT\n",
-            iptables()
-        ));
-        s.push_str(&format!(
-            "{} -A SHROUD_KILLSWITCH -o tap+ -j ACCEPT\n",
-            iptables()
-        ));
-        s.push_str(&format!(
-            "{} -A SHROUD_KILLSWITCH -o wg+ -j ACCEPT\n",
-            iptables()
-        ));
-
-        s.push_str(&self.build_doh_blocking_rules());
-
-        for ip in vpn_ips {
-            if let IpAddr::V4(v4) = ip {
-                s.push_str(&format!(
-                    "{} -A SHROUD_KILLSWITCH -d {} -j ACCEPT\n",
-                    iptables(),
-                    v4
-                ));
-            }
-        }
-
-        // LAN access rules — use detected subnets instead of full RFC1918
-        let lan_subnets = crate::killswitch::rules::detect_local_subnets();
-        for subnet in &lan_subnets {
-            // SECURITY: Double-check that detected subnets are valid private CIDRs.
-            // Rejects 0.0.0.0/0 or public ranges that would open the kill switch
-            // to all traffic (SHROUD-VULN-021).
-            if !crate::killswitch::rules::is_valid_private_cidr(subnet) {
-                warn!("Rejected non-private subnet in iptables rules: {}", subnet);
-                continue;
-            }
-            s.push_str(&format!(
-                "{} -A SHROUD_KILLSWITCH -d {} -j ACCEPT\n",
-                iptables(),
-                subnet
-            ));
-        }
-        s.push_str(&format!(
-            "{} -A SHROUD_KILLSWITCH -p udp --dport 67 -j ACCEPT\n",
-            iptables()
-        ));
-        s.push_str(&format!(
-            "{} -A SHROUD_KILLSWITCH -p udp --sport 68 -j ACCEPT\n",
-            iptables()
-        ));
-
-        s.push_str(&format!(
-            "{} -A SHROUD_KILLSWITCH -m limit --limit 1/sec -j LOG --log-prefix SHROUD-KS-DROP --log-level 4\n",
-            iptables()
-        ));
-        s.push_str(&format!("{} -A SHROUD_KILLSWITCH -j DROP\n", iptables()));
-
-        // SECURITY: Insert jump rule LAST — chain is now fully populated,
-        // so traffic is never directed to an incomplete rule chain.
-        s.push_str(&format!(
-            "{} -I OUTPUT 1 -j SHROUD_KILLSWITCH\n",
-            iptables()
-        ));
-
-        // IPv6
-        match self.ipv6_mode {
-            Ipv6Mode::Block => {
-                s.push_str(&format!(
-                    "{} -I OUTPUT 1 -o lo -j ACCEPT 2>/dev/null || true\n",
-                    ip6tables()
-                ));
-                s.push_str(&format!(
-                    "{} -I OUTPUT 2 -j DROP 2>/dev/null || true\n",
-                    ip6tables()
-                ));
-            }
-            Ipv6Mode::Tunnel => {
-                s.push_str(&format!(
-                    "{} -I OUTPUT 1 -o lo -j ACCEPT 2>/dev/null || true\n",
-                    ip6tables()
-                ));
-                s.push_str(&format!(
-                    "{} -I OUTPUT 2 -m conntrack --ctstate ESTABLISHED,RELATED -j ACCEPT 2>/dev/null || true\n",
-                    ip6tables()
-                ));
-                s.push_str(&format!(
-                    "{} -I OUTPUT 3 -o tun+ -j ACCEPT 2>/dev/null || true\n",
-                    ip6tables()
-                ));
-                s.push_str(&format!(
-                    "{} -I OUTPUT 4 -d fe80::/10 -j ACCEPT 2>/dev/null || true\n",
-                    ip6tables()
-                ));
-                s.push_str(&format!(
-                    "{} -I OUTPUT 5 -j DROP 2>/dev/null || true\n",
-                    ip6tables()
-                ));
-            }
-            Ipv6Mode::Off => {}
-        }
-
-        s
-    }
-
-    /// Build a preview of the kill switch rules script (for diagnostics/tests)
-    #[allow(dead_code)]
-    pub fn build_rules_preview(&self, vpn_ips: &[IpAddr]) -> String {
-        self.build_complete_script(vpn_ips)
-    }
-
-    fn build_dns_rules(&self) -> String {
-        let mut s = String::new();
-
-        match self.dns_mode {
-            DnsMode::Tunnel | DnsMode::Strict => {
-                s.push_str("# DNS Leak Protection (Tunnel/Strict)\n");
-                s.push_str(&format!(
-                    "{} -A SHROUD_KILLSWITCH -o tun+ -p udp --dport 53 -j ACCEPT\n",
-                    iptables()
-                ));
-                s.push_str(&format!(
-                    "{} -A SHROUD_KILLSWITCH -o tun+ -p tcp --dport 53 -j ACCEPT\n",
-                    iptables()
-                ));
-                s.push_str(&format!(
-                    "{} -A SHROUD_KILLSWITCH -o wg+ -p udp --dport 53 -j ACCEPT\n",
-                    iptables()
-                ));
-                s.push_str(&format!(
-                    "{} -A SHROUD_KILLSWITCH -o wg+ -p tcp --dport 53 -j ACCEPT\n",
-                    iptables()
-                ));
-                s.push_str(&format!(
-                    "{} -A SHROUD_KILLSWITCH -o tap+ -p udp --dport 53 -j ACCEPT\n",
-                    iptables()
-                ));
-                s.push_str(&format!(
-                    "{} -A SHROUD_KILLSWITCH -o tap+ -p tcp --dport 53 -j ACCEPT\n",
-                    iptables()
-                ));
-                s.push_str(&format!(
-                    "{} -A SHROUD_KILLSWITCH -p udp --dport 53 -j DROP\n",
-                    iptables()
-                ));
-                s.push_str(&format!(
-                    "{} -A SHROUD_KILLSWITCH -p tcp --dport 53 -j DROP\n",
-                    iptables()
-                ));
-                s.push_str(&format!(
-                    "{} -A SHROUD_KILLSWITCH -p tcp --dport 853 -j DROP\n",
-                    iptables()
-                ));
-            }
-            DnsMode::Localhost => {
-                s.push_str("# DNS Leak Protection (Localhost)\n");
-                // SECURITY: Only allow DNS to 127.0.0.1 and 127.0.0.53 (systemd-resolved),
-                // not the entire 127.0.0.0/8 range, to prevent rogue resolvers on other
-                // loopback addresses.
-                s.push_str(&format!(
-                    "{} -A SHROUD_KILLSWITCH -d 127.0.0.1 -p udp --dport 53 -j ACCEPT\n",
-                    iptables()
-                ));
-                s.push_str(&format!(
-                    "{} -A SHROUD_KILLSWITCH -d 127.0.0.1 -p tcp --dport 53 -j ACCEPT\n",
-                    iptables()
-                ));
-                s.push_str(&format!(
-                    "{} -A SHROUD_KILLSWITCH -d 127.0.0.53 -p udp --dport 53 -j ACCEPT\n",
-                    iptables()
-                ));
-                s.push_str(&format!(
-                    "{} -A SHROUD_KILLSWITCH -d 127.0.0.53 -p tcp --dport 53 -j ACCEPT\n",
-                    iptables()
-                ));
-                s.push_str(&format!(
-                    "{} -A SHROUD_KILLSWITCH -d ::1 -p udp --dport 53 -j ACCEPT\n",
-                    iptables()
-                ));
-                s.push_str(&format!(
-                    "{} -A SHROUD_KILLSWITCH -d ::1 -p tcp --dport 53 -j ACCEPT\n",
-                    iptables()
-                ));
-                s.push_str(&format!(
-                    "{} -A SHROUD_KILLSWITCH -p udp --dport 53 -j DROP\n",
-                    iptables()
-                ));
-                s.push_str(&format!(
-                    "{} -A SHROUD_KILLSWITCH -p tcp --dport 53 -j DROP\n",
-                    iptables()
-                ));
-                s.push_str(&format!(
-                    "{} -A SHROUD_KILLSWITCH -p tcp --dport 853 -j DROP\n",
-                    iptables()
-                ));
-            }
-            DnsMode::Any => {
-                s.push_str("# DNS (Any Mode - NOT RECOMMENDED)\n");
-                s.push_str(&format!(
-                    "{} -A SHROUD_KILLSWITCH -p udp --dport 53 -j ACCEPT\n",
-                    iptables()
-                ));
-                s.push_str(&format!(
-                    "{} -A SHROUD_KILLSWITCH -p tcp --dport 53 -j ACCEPT\n",
-                    iptables()
-                ));
-            }
-        }
-
-        s
-    }
-
-    fn build_doh_blocking_rules(&self) -> String {
-        if !self.block_doh {
-            return String::new();
-        }
-
-        if !matches!(self.dns_mode, DnsMode::Tunnel | DnsMode::Strict) {
-            return String::new();
-        }
-
-        let mut s = String::new();
-        s.push_str("# Block DNS-over-HTTPS (DoH) to known providers\n");
-
-        for ip in DOH_PROVIDER_IPS
-            .iter()
-            .copied()
-            .chain(self.custom_doh_blocklist.iter().map(|s| s.as_str()))
-        {
-            // SECURITY: Validate each IP to prevent injection into iptables
-            // commands. custom_doh_blocklist comes from config.toml and could
-            // contain crafted strings (SHROUD-VULN-022).
-            if !crate::killswitch::rules::is_valid_ipv4(ip) {
-                warn!(
-                    "Rejected invalid DoH blocklist IP (possible injection): {}",
-                    ip
-                );
-                continue;
-            }
-            s.push_str(&format!(
-                "{} -A SHROUD_KILLSWITCH -d {} -p tcp --dport 443 -j DROP\n",
-                iptables(),
-                ip
-            ));
-        }
-
-        s
-    }
-
-    /// Timeout for sudo/iptables commands (seconds)
-    /// Long enough for normal operations, short enough to detect hangs
-    const SUDO_CMD_TIMEOUT_SECS: u64 = 30;
-
-    async fn run_single_script(&self, script: &str) -> Result<(), KillSwitchError> {
-        use tokio::process::Command;
-        use tokio::time::{timeout, Duration};
-
-        for raw_line in script.lines() {
-            let mut line = raw_line.trim().to_string();
-            if line.is_empty() || line.starts_with('#') {
-                continue;
-            }
-
-            let ignore_error = line.contains("|| true");
-            if let Some(stripped) = line.strip_suffix("|| true") {
-                line = stripped.trim().to_string();
-            }
-            line = line.replace("2>/dev/null", "").trim().to_string();
-            if line.is_empty() || line == "exit 0" {
-                continue;
-            }
-
-            let parts: Vec<&str> = line.split_whitespace().collect();
-            if parts.is_empty() {
-                continue;
-            }
-
-            let mut cmd = parts[0];
-            if self.use_legacy
-                && (parts[0].ends_with("iptables") || parts[0].ends_with("ip6tables"))
-            {
-                if let Some(legacy_cmd) = Self::legacy_variant(parts[0]).await {
-                    cmd = legacy_cmd;
-                }
-            }
-
-            // HARDENING: Add timeout to sudo commands to prevent hanging
-            // on password prompts or frozen iptables kernel modules
-            let output = match timeout(
-                Duration::from_secs(Self::SUDO_CMD_TIMEOUT_SECS),
-                Command::new("sudo")
-                    .arg("-n") // non-interactive: fail immediately if password needed
-                    .arg(cmd)
-                    .args(&parts[1..])
-                    .output(),
-            )
-            .await
-            {
-                Ok(Ok(output)) => output,
-                Ok(Err(e)) => return Err(KillSwitchError::Spawn(e)),
-                Err(_) => {
-                    return Err(KillSwitchError::Command(format!(
-                        "Command timed out after {}s (possible sudo prompt or frozen kernel module): {}",
-                        Self::SUDO_CMD_TIMEOUT_SECS,
-                        line
-                    )));
-                }
-            };
-
-            if !output.status.success() && !ignore_error {
-                let code = output.status.code().unwrap_or(-1);
-                if code == 126 || code == 127 {
-                    return Err(KillSwitchError::Permission);
-                }
-                let stderr = String::from_utf8_lossy(&output.stderr);
-                let stderr_lower = stderr.to_lowercase();
-
-                if (stderr_lower.contains("cache initialization failed")
-                    || stderr_lower.contains("netlink: error")
-                    || stderr_lower.contains("can't initialize iptables table")
-                    || stderr_lower.contains("ip_tables"))
-                    && !parts.is_empty()
-                    && (parts[0].ends_with("iptables") || parts[0].ends_with("ip6tables"))
-                {
-                    if let Some(legacy_cmd) = Self::legacy_variant(parts[0]).await {
-                        // HARDENING: Add timeout to legacy fallback as well
-                        let legacy_output = match timeout(
-                            Duration::from_secs(Self::SUDO_CMD_TIMEOUT_SECS),
-                            Command::new("sudo")
-                                .arg("-n")
-                                .arg(legacy_cmd)
-                                .args(&parts[1..])
-                                .output(),
-                        )
-                        .await
-                        {
-                            Ok(Ok(output)) => output,
-                            Ok(Err(e)) => return Err(KillSwitchError::Spawn(e)),
-                            Err(_) => {
-                                return Err(KillSwitchError::Command(format!(
-                                    "Legacy command timed out after {}s: {}",
-                                    Self::SUDO_CMD_TIMEOUT_SECS,
-                                    line
-                                )));
-                            }
-                        };
-
-                        if legacy_output.status.success() {
-                            continue;
-                        }
-                    }
-
-                    let detail = if stderr.trim().is_empty() {
-                        line.clone()
-                    } else {
-                        stderr.trim().to_string()
-                    };
-
-                    return Err(KillSwitchError::Command(format!(
-                        "{} (iptables-nft failed; install iptables-legacy or nftables)",
-                        detail
-                    )));
-                }
-
-                let detail = if stderr.trim().is_empty() {
-                    line.clone()
-                } else {
-                    stderr.trim().to_string()
-                };
-                return Err(KillSwitchError::Command(format!(
-                    "Command failed (exit {}): {}",
-                    code, detail
-                )));
-            }
-        }
-
-        Ok(())
-    }
-
-    async fn legacy_variant(cmd: &str) -> Option<&'static str> {
-        let (candidate, candidate_path) = if cmd.ends_with("iptables") {
-            ("iptables-legacy", "/usr/sbin/iptables-legacy")
-        } else if cmd.ends_with("ip6tables") {
-            ("ip6tables-legacy", "/usr/sbin/ip6tables-legacy")
-        } else {
-            return None;
-        };
-
-        if Command::new(candidate)
-            .arg("--version")
-            .stdout(Stdio::null())
-            .stderr(Stdio::null())
-            .status()
-            .await
-            .map(|s| s.success())
-            .unwrap_or(false)
-        {
-            return Some(candidate);
-        }
-
-        if Command::new(candidate_path)
-            .arg("--version")
-            .stdout(Stdio::null())
-            .stderr(Stdio::null())
-            .status()
-            .await
-            .map(|s| s.success())
-            .unwrap_or(false)
-        {
-            return Some(candidate_path);
-        }
-
-        None
-    }
-
     async fn select_backend(&mut self) -> Result<FirewallBackend, KillSwitchError> {
         // Prefer nftables: atomic rule application means no traffic gap during updates.
         // Fall back to iptables/iptables-legacy if nft is unavailable.
@@ -896,118 +435,6 @@ impl KillSwitch {
             KillSwitchError::Spawn(_) | KillSwitchError::NotFound => true,
             _ => false,
         }
-    }
-
-    async fn nft_is_available() -> bool {
-        if Command::new("nft")
-            .arg("--version")
-            .stdout(Stdio::null())
-            .stderr(Stdio::null())
-            .status()
-            .await
-            .map(|s| s.success())
-            .unwrap_or(false)
-        {
-            return true;
-        }
-
-        if Command::new(nft())
-            .arg("--version")
-            .stdout(Stdio::null())
-            .stderr(Stdio::null())
-            .status()
-            .await
-            .map(|s| s.success())
-            .unwrap_or(false)
-        {
-            return true;
-        }
-
-        Command::new("sudo")
-            .args(["-n", nft(), "--version"])
-            .stdout(Stdio::null())
-            .stderr(Stdio::null())
-            .status()
-            .await
-            .map(|s| s.success())
-            .unwrap_or(false)
-    }
-
-    /// Check if sudo is available and configured for passwordless iptables.
-    pub fn check_sudo_access() -> Result<(), KillSwitchError> {
-        let output = std::process::Command::new("sudo")
-            .args(["-n", iptables(), "-L", "-n"])
-            .stdin(Stdio::null())
-            .stdout(Stdio::null())
-            .stderr(Stdio::piped())
-            .output()
-            .map_err(KillSwitchError::Spawn)?;
-
-        let stderr = String::from_utf8_lossy(&output.stderr);
-        let stderr_lower = stderr.to_lowercase();
-        if stderr_lower.contains("ip_tables")
-            || stderr_lower.contains("table does not exist")
-            || stderr_lower.contains("can't initialize iptables table")
-            || stderr_lower.contains("cache initialization failed")
-            || stderr_lower.contains("netlink: error")
-        {
-            return Err(KillSwitchError::Command(format!(
-                "Sudo check failed: {}",
-                stderr.trim()
-            )));
-        }
-
-        if output.status.success() {
-            return Ok(());
-        }
-
-        if stderr_lower.contains("permission denied") || stderr_lower.contains("password") {
-            return Err(KillSwitchError::Permission);
-        }
-
-        Err(KillSwitchError::Command(format!(
-            "Sudo check failed: {}",
-            stderr.trim()
-        )))
-    }
-
-    fn check_iptables_legacy_access() -> Result<bool, KillSwitchError> {
-        let output = std::process::Command::new("sudo")
-            .args(["-n", "iptables-legacy", "-L", "-n"])
-            .stdin(Stdio::null())
-            .stdout(Stdio::null())
-            .stderr(Stdio::piped())
-            .output();
-
-        match output {
-            Ok(out) => Ok(out.status.success()),
-            Err(_) => Ok(false),
-        }
-    }
-
-    /// Verify our rules are actually in place
-    async fn verify_rules_exist(&self) -> bool {
-        // Check if our chain exists and has the jump rule
-        // Use sudo -n to avoid password prompts
-        let output = Command::new("sudo")
-            .args(["-n", iptables(), "-C", "OUTPUT", "-j", CHAIN_NAME])
-            .stdout(Stdio::null())
-            .stderr(Stdio::null())
-            .status()
-            .await;
-
-        matches!(output, Ok(status) if status.success())
-    }
-
-    async fn verify_nft_rules_exist(&self) -> bool {
-        let output = Command::new("sudo")
-            .args(["-n", nft(), "list", "table", "inet", NFT_TABLE])
-            .stdout(Stdio::null())
-            .stderr(Stdio::null())
-            .status()
-            .await;
-
-        matches!(output, Ok(status) if status.success())
     }
 
     /// Disable the kill switch.
@@ -1090,111 +517,6 @@ impl KillSwitch {
         Ok(())
     }
 
-    /// Robust cleanup that removes ALL duplicate rules (handles race conditions)
-    async fn robust_iptables_cleanup(&self) {
-        // Remove ALL duplicate SHROUD_KILLSWITCH jump rules from OUTPUT chain
-        for _ in 0..100 {
-            // Safety limit
-            let output = Command::new("sudo")
-                .args(["-n", iptables(), "-D", "OUTPUT", "-j", CHAIN_NAME])
-                .stdout(Stdio::null())
-                .stderr(Stdio::null())
-                .status()
-                .await;
-            if !matches!(output, Ok(s) if s.success()) {
-                break;
-            }
-        }
-
-        // Flush and delete the chain
-        let _ = Command::new("sudo")
-            .args(["-n", iptables(), "-F", CHAIN_NAME])
-            .stdout(Stdio::null())
-            .stderr(Stdio::null())
-            .status()
-            .await;
-
-        let _ = Command::new("sudo")
-            .args(["-n", iptables(), "-X", CHAIN_NAME])
-            .stdout(Stdio::null())
-            .stderr(Stdio::null())
-            .status()
-            .await;
-
-        // IPv6: Remove ALL duplicate rules inserted by Block/Tunnel modes
-        // These are direct OUTPUT rules, not a chain
-        let ipv6_rules_to_remove = [
-            vec!["-D", "OUTPUT", "-j", "DROP"],
-            vec!["-D", "OUTPUT", "-o", "lo", "-j", "ACCEPT"],
-            vec![
-                "-D",
-                "OUTPUT",
-                "-m",
-                "conntrack",
-                "--ctstate",
-                "ESTABLISHED,RELATED",
-                "-j",
-                "ACCEPT",
-            ],
-            vec!["-D", "OUTPUT", "-o", "tun+", "-j", "ACCEPT"],
-            vec!["-D", "OUTPUT", "-d", "fe80::/10", "-j", "ACCEPT"],
-        ];
-
-        for rule in &ipv6_rules_to_remove {
-            // Remove ALL duplicates of each rule
-            for _ in 0..100 {
-                let mut cmd = Command::new("sudo");
-                cmd.arg("-n").arg(ip6tables());
-                for arg in rule {
-                    cmd.arg(arg);
-                }
-                let output = cmd
-                    .stdout(Stdio::null())
-                    .stderr(Stdio::null())
-                    .status()
-                    .await;
-                if !matches!(output, Ok(s) if s.success()) {
-                    break;
-                }
-            }
-        }
-
-        // Also try to clean up IPv6 SHROUD_KILLSWITCH chain if it exists
-        for _ in 0..100 {
-            let output = Command::new("sudo")
-                .args(["-n", ip6tables(), "-D", "OUTPUT", "-j", CHAIN_NAME])
-                .stdout(Stdio::null())
-                .stderr(Stdio::null())
-                .status()
-                .await;
-            if !matches!(output, Ok(s) if s.success()) {
-                break;
-            }
-        }
-
-        let _ = Command::new("sudo")
-            .args(["-n", ip6tables(), "-F", CHAIN_NAME])
-            .stdout(Stdio::null())
-            .stderr(Stdio::null())
-            .status()
-            .await;
-
-        let _ = Command::new("sudo")
-            .args(["-n", ip6tables(), "-X", CHAIN_NAME])
-            .stdout(Stdio::null())
-            .stderr(Stdio::null())
-            .status()
-            .await;
-
-        // Clean up nftables table too
-        let _ = Command::new("sudo")
-            .args(["-n", nft(), "delete", "table", "inet", NFT_TABLE])
-            .stdout(Stdio::null())
-            .stderr(Stdio::null())
-            .status()
-            .await;
-    }
-
     /// Update the kill switch rules (e.g., when VPN interface changes)
     #[allow(dead_code)]
     pub async fn update(&mut self) -> Result<(), KillSwitchError> {
@@ -1212,290 +534,6 @@ impl KillSwitch {
                 self.enable_nft(&vpn_server_ips).await
             }
         }
-    }
-
-    async fn enable_nft(&mut self, vpn_server_ips: &[IpAddr]) -> Result<(), KillSwitchError> {
-        let _ = self.cleanup_nft_table().await;
-        let rules = self.build_nft_ruleset(vpn_server_ips);
-        self.run_nft(&["-f", "-"], Some(&rules)).await?;
-        self.enabled = true;
-        Ok(())
-    }
-
-    async fn disable_nft(&self) -> Result<(), KillSwitchError> {
-        self.cleanup_nft_table().await
-    }
-
-    async fn cleanup_nft_table(&self) -> Result<(), KillSwitchError> {
-        let result = self
-            .run_nft(&["delete", "table", "inet", NFT_TABLE], None)
-            .await;
-
-        match result {
-            Ok(_) => Ok(()),
-            Err(KillSwitchError::Command(msg))
-                if msg.contains("No such file") || msg.contains("does not exist") =>
-            {
-                Ok(())
-            }
-            Err(e) => Err(e),
-        }
-    }
-
-    /// Timeout for nft commands (same as iptables)
-    const NFT_CMD_TIMEOUT_SECS: u64 = 30;
-
-    async fn run_nft(
-        &self,
-        args: &[&str],
-        stdin_data: Option<&str>,
-    ) -> Result<(), KillSwitchError> {
-        use std::time::Duration;
-        use tokio::time::timeout;
-
-        let mut cmd = Command::new("sudo");
-        cmd.arg("-n"); // Non-interactive to avoid password prompts
-        cmd.arg(nft());
-        cmd.args(args);
-
-        if stdin_data.is_some() {
-            cmd.stdin(Stdio::piped());
-        }
-        cmd.stdout(Stdio::piped());
-        cmd.stderr(Stdio::piped());
-
-        let result = timeout(Duration::from_secs(Self::NFT_CMD_TIMEOUT_SECS), async {
-            let mut child = cmd.spawn().map_err(KillSwitchError::Spawn)?;
-
-            if let Some(data) = stdin_data {
-                use tokio::io::AsyncWriteExt;
-                if let Some(mut stdin) = child.stdin.take() {
-                    stdin
-                        .write_all(data.as_bytes())
-                        .await
-                        .map_err(KillSwitchError::Write)?;
-                }
-            }
-
-            child
-                .wait_with_output()
-                .await
-                .map_err(KillSwitchError::Wait)
-        })
-        .await;
-
-        match result {
-            Ok(Ok(output)) => {
-                if output.status.success() {
-                    Ok(())
-                } else {
-                    let stderr = String::from_utf8_lossy(&output.stderr);
-                    Err(KillSwitchError::Command(stderr.trim().to_string()))
-                }
-            }
-            Ok(Err(e)) => Err(e),
-            Err(_) => {
-                warn!(
-                    "nft command timed out after {}s",
-                    Self::NFT_CMD_TIMEOUT_SECS
-                );
-                Err(KillSwitchError::Command(format!(
-                    "nft command timed out after {}s",
-                    Self::NFT_CMD_TIMEOUT_SECS
-                )))
-            }
-        }
-    }
-
-    fn build_nft_ruleset(&self, vpn_server_ips: &[IpAddr]) -> String {
-        let mut rules = format!(
-            r#"
-table inet {table} {{
-    chain output {{
-        type filter hook output priority 0; policy drop;
-
-        # === LOOPBACK ===
-        oifname "lo" accept
-
-        # === ESTABLISHED/RELATED ===
-        ct state established,related accept
-"#,
-            table = NFT_TABLE
-        );
-
-        match self.ipv6_mode {
-            Ipv6Mode::Block => {
-                rules.push_str(
-                    r#"
-        # === IPv6 LEAK PROTECTION (block mode) ===
-        meta nfproto ipv6 drop
-"#,
-                );
-            }
-            Ipv6Mode::Tunnel => {
-                rules.push_str(
-                    r#"
-        # === IPv6 LEAK PROTECTION (tunnel mode) ===
-        ip6 daddr fe80::/10 accept
-"#,
-                );
-            }
-            Ipv6Mode::Off => {
-                rules.push_str(
-                    r#"
-        # === IPv6 (off - no special handling) ===
-        ip6 daddr fe80::/10 accept
-"#,
-                );
-            }
-        }
-
-        rules.push_str(
-            r#"
-        # === DHCP ===
-        udp dport 67 accept
-        udp sport 68 accept
-"#,
-        );
-
-        match self.dns_mode {
-            DnsMode::Tunnel | DnsMode::Strict => {
-                rules.push_str(
-                    r#"
-        # === DNS LEAK PROTECTION (tunnel/strict) ===
-        oifname "tun*" udp dport 53 accept
-        oifname "tun*" tcp dport 53 accept
-        oifname "wg*" udp dport 53 accept
-        oifname "wg*" tcp dport 53 accept
-        oifname "tap*" udp dport 53 accept
-        oifname "tap*" tcp dport 53 accept
-        udp dport 53 drop
-        tcp dport 53 drop
-        tcp dport 853 drop
-"#,
-                );
-            }
-            DnsMode::Localhost => {
-                rules.push_str(
-                    r#"
-        # === DNS LEAK PROTECTION (localhost) ===
-        ip daddr 127.0.0.1 udp dport 53 accept
-        ip daddr 127.0.0.1 tcp dport 53 accept
-        ip daddr 127.0.0.53 udp dport 53 accept
-        ip daddr 127.0.0.53 tcp dport 53 accept
-"#,
-                );
-                if self.ipv6_mode != Ipv6Mode::Block {
-                    rules.push_str(
-                        r#"
-        ip6 daddr ::1 udp dport 53 accept
-        ip6 daddr ::1 tcp dport 53 accept
-"#,
-                    );
-                }
-                rules.push_str(
-                    r#"
-        udp dport 53 drop
-        tcp dport 53 drop
-        tcp dport 853 drop
-"#,
-                );
-            }
-            DnsMode::Any => {
-                rules.push_str(
-                    r#"
-        # === DNS (any mode - LEGACY/INSECURE) ===
-        udp dport 53 accept
-        tcp dport 53 accept
-"#,
-                );
-            }
-        }
-
-        if self.block_doh && matches!(self.dns_mode, DnsMode::Tunnel | DnsMode::Strict) {
-            rules.push_str("\n        # === Block DNS-over-HTTPS (DoH) ===\n");
-            for ip in DOH_PROVIDER_IPS
-                .iter()
-                .copied()
-                .chain(self.custom_doh_blocklist.iter().map(|s| s.as_str()))
-            {
-                // SECURITY: Validate each IP to prevent nft ruleset injection.
-                // custom_doh_blocklist comes from config.toml (SHROUD-VULN-022).
-                if !crate::killswitch::rules::is_valid_ipv4(ip) {
-                    warn!(
-                        "Rejected invalid DoH blocklist IP in nft (possible injection): {}",
-                        ip
-                    );
-                    continue;
-                }
-                rules.push_str(&format!("        ip daddr {} tcp dport 443 drop\n", ip));
-            }
-        }
-
-        // === LOCAL NETWORK (auto-detected subnets) ===
-        let lan_subnets = crate::killswitch::rules::detect_local_subnets();
-        rules.push_str("\n        # === LOCAL NETWORK ===\n");
-        for subnet in &lan_subnets {
-            // SECURITY: Double-check that detected subnets are valid private CIDRs
-            if !crate::killswitch::rules::is_valid_private_cidr(subnet) {
-                warn!("Rejected non-private subnet in nft rules: {}", subnet);
-                continue;
-            }
-            rules.push_str(&format!("        ip daddr {} accept\n", subnet));
-        }
-
-        rules.push_str(
-            r#"
-        # === VPN TUNNEL INTERFACES ===
-        oifname "tun*" accept
-        oifname "tap*" accept
-        oifname "wg*" accept
-"#,
-        );
-
-        if !vpn_server_ips.is_empty() {
-            rules.push_str("\n        # === VPN SERVER ALLOWLIST ===\n");
-        }
-        for ip in vpn_server_ips {
-            match ip {
-                IpAddr::V4(v4) => {
-                    rules.push_str(&format!("        ip daddr {} accept\n", v4));
-                }
-                IpAddr::V6(v6) => {
-                    if self.ipv6_mode != Ipv6Mode::Block {
-                        rules.push_str(&format!("        ip6 daddr {} accept\n", v6));
-                    }
-                }
-            }
-        }
-
-        if let Some(ip) = self.vpn_server_ip {
-            match ip {
-                IpAddr::V4(v4) => {
-                    rules.push_str(&format!("        ip daddr {} accept\n", v4));
-                }
-                IpAddr::V6(v6) => {
-                    if self.ipv6_mode != Ipv6Mode::Block {
-                        rules.push_str(&format!("        ip6 daddr {} accept\n", v6));
-                    }
-                }
-            }
-        }
-
-        rules.push_str(
-            r#"
-        # === DEFAULT DROP ===
-        limit rate 1/second log prefix "SHROUD-KS-DROP" drop
-    }
-
-    chain input {
-        type filter hook input priority 0; policy accept;
-    }
-}
-"#,
-        );
-
-        rules
     }
 
     /// Detect the VPN tunnel interface from the system
@@ -2092,203 +1130,14 @@ mod security_tests {
 }
 
 // =========================================================================
-// nftables ruleset tests
-// =========================================================================
-
-#[cfg(test)]
-mod nft_tests {
-    use super::*;
-
-    #[test]
-    fn test_nft_ruleset_basic_structure() {
-        let ks = KillSwitch::new();
-        let rules = ks.build_nft_ruleset(&[]);
-        assert!(rules.contains("table inet shroud_killswitch"));
-        assert!(rules.contains("chain output"));
-        assert!(rules.contains("policy drop"));
-        assert!(rules.contains("oifname \"lo\" accept"));
-        assert!(rules.contains("ct state established,related accept"));
-    }
-
-    #[test]
-    fn test_nft_ruleset_vpn_interfaces() {
-        let ks = KillSwitch::new();
-        let rules = ks.build_nft_ruleset(&[]);
-        assert!(rules.contains("oifname \"tun*\" accept"));
-        assert!(rules.contains("oifname \"tap*\" accept"));
-        assert!(rules.contains("oifname \"wg*\" accept"));
-    }
-
-    #[test]
-    fn test_nft_ruleset_local_network() {
-        let ks = KillSwitch::new();
-        let rules = ks.build_nft_ruleset(&[]);
-        // LAN rules now use auto-detected subnets (or RFC1918 fallback)
-        assert!(
-            rules.contains("ip daddr")
-                && (rules.contains("192.168.")
-                    || rules.contains("10.0.")
-                    || rules.contains("172.16.")
-                    || rules.contains("169.254.")),
-            "NFT ruleset should contain LAN subnet accept rules"
-        );
-    }
-
-    #[test]
-    fn test_nft_ruleset_dhcp() {
-        let ks = KillSwitch::new();
-        let rules = ks.build_nft_ruleset(&[]);
-        assert!(rules.contains("udp dport 67 accept"));
-        assert!(rules.contains("udp sport 68 accept"));
-    }
-
-    #[test]
-    fn test_nft_ruleset_ipv6_block() {
-        let ks = KillSwitch::with_config(DnsMode::Tunnel, Ipv6Mode::Block, true, Vec::new());
-        let rules = ks.build_nft_ruleset(&[]);
-        assert!(rules.contains("meta nfproto ipv6 drop"));
-    }
-
-    #[test]
-    fn test_nft_ruleset_ipv6_tunnel() {
-        let ks = KillSwitch::with_config(DnsMode::Tunnel, Ipv6Mode::Tunnel, true, Vec::new());
-        let rules = ks.build_nft_ruleset(&[]);
-        assert!(rules.contains("fe80::/10 accept"));
-        assert!(!rules.contains("meta nfproto ipv6 drop"));
-    }
-
-    #[test]
-    fn test_nft_ruleset_ipv6_off() {
-        let ks = KillSwitch::with_config(DnsMode::Tunnel, Ipv6Mode::Off, true, Vec::new());
-        let rules = ks.build_nft_ruleset(&[]);
-        assert!(!rules.contains("meta nfproto ipv6 drop"));
-        assert!(rules.contains("fe80::/10 accept"));
-    }
-
-    #[test]
-    fn test_nft_dns_tunnel_mode() {
-        let ks = KillSwitch::with_config(DnsMode::Tunnel, Ipv6Mode::Block, true, Vec::new());
-        let rules = ks.build_nft_ruleset(&[]);
-        assert!(rules.contains("oifname \"tun*\" udp dport 53 accept"));
-        assert!(rules.contains("oifname \"wg*\" udp dport 53 accept"));
-        assert!(rules.contains("udp dport 53 drop"));
-        assert!(rules.contains("tcp dport 53 drop"));
-        assert!(rules.contains("tcp dport 853 drop"));
-    }
-
-    #[test]
-    fn test_nft_dns_localhost_mode() {
-        let ks = KillSwitch::with_config(DnsMode::Localhost, Ipv6Mode::Off, true, Vec::new());
-        let rules = ks.build_nft_ruleset(&[]);
-        assert!(rules.contains("ip daddr 127.0.0.1 udp dport 53 accept"));
-        assert!(rules.contains("ip daddr 127.0.0.53 udp dport 53 accept"));
-        assert!(rules.contains("udp dport 53 drop"));
-    }
-
-    #[test]
-    fn test_nft_dns_localhost_with_ipv6() {
-        let ks = KillSwitch::with_config(DnsMode::Localhost, Ipv6Mode::Tunnel, true, Vec::new());
-        let rules = ks.build_nft_ruleset(&[]);
-        assert!(rules.contains("ip6 daddr ::1 udp dport 53 accept"));
-    }
-
-    #[test]
-    fn test_nft_dns_localhost_without_ipv6() {
-        let ks = KillSwitch::with_config(DnsMode::Localhost, Ipv6Mode::Block, true, Vec::new());
-        let rules = ks.build_nft_ruleset(&[]);
-        assert!(!rules.contains("ip6 daddr ::1"));
-    }
-
-    #[test]
-    fn test_nft_dns_any_mode() {
-        let ks = KillSwitch::with_config(DnsMode::Any, Ipv6Mode::Block, false, Vec::new());
-        let rules = ks.build_nft_ruleset(&[]);
-        assert!(rules.contains("udp dport 53 accept"));
-        assert!(rules.contains("tcp dport 53 accept"));
-        assert!(!rules.contains("udp dport 53 drop"));
-    }
-
-    #[test]
-    fn test_nft_doh_blocking() {
-        let ks = KillSwitch::with_config(DnsMode::Strict, Ipv6Mode::Block, true, Vec::new());
-        let rules = ks.build_nft_ruleset(&[]);
-        assert!(rules.contains("ip daddr 1.1.1.1 tcp dport 443 drop"));
-        assert!(rules.contains("ip daddr 8.8.8.8 tcp dport 443 drop"));
-        assert!(rules.contains("ip daddr 9.9.9.9 tcp dport 443 drop"));
-    }
-
-    #[test]
-    fn test_nft_doh_blocking_disabled() {
-        let ks = KillSwitch::with_config(DnsMode::Tunnel, Ipv6Mode::Block, false, Vec::new());
-        let rules = ks.build_nft_ruleset(&[]);
-        assert!(!rules.contains("tcp dport 443 drop"));
-    }
-
-    #[test]
-    fn test_nft_custom_doh_blocklist() {
-        let custom = vec!["100.100.100.100".into(), "200.200.200.200".into()];
-        let ks = KillSwitch::with_config(DnsMode::Strict, Ipv6Mode::Block, true, custom);
-        let rules = ks.build_nft_ruleset(&[]);
-        assert!(rules.contains("ip daddr 100.100.100.100 tcp dport 443 drop"));
-        assert!(rules.contains("ip daddr 200.200.200.200 tcp dport 443 drop"));
-    }
-
-    #[test]
-    fn test_nft_vpn_server_ips() {
-        let ks = KillSwitch::new();
-        let server_ip: IpAddr = "203.0.113.50".parse().unwrap();
-        let rules = ks.build_nft_ruleset(&[server_ip]);
-        assert!(rules.contains("ip daddr 203.0.113.50 accept"));
-    }
-
-    #[test]
-    fn test_nft_vpn_server_ipv6() {
-        let ks = KillSwitch::with_config(DnsMode::Tunnel, Ipv6Mode::Tunnel, true, Vec::new());
-        let server_ip: IpAddr = "2001:db8::1".parse().unwrap();
-        let rules = ks.build_nft_ruleset(&[server_ip]);
-        assert!(rules.contains("ip6 daddr 2001:db8::1 accept"));
-    }
-
-    #[test]
-    fn test_nft_vpn_server_ipv6_blocked_when_block_mode() {
-        let ks = KillSwitch::with_config(DnsMode::Tunnel, Ipv6Mode::Block, true, Vec::new());
-        let server_ip: IpAddr = "2001:db8::1".parse().unwrap();
-        let rules = ks.build_nft_ruleset(&[server_ip]);
-        assert!(!rules.contains("ip6 daddr 2001:db8::1 accept"));
-    }
-
-    #[test]
-    fn test_nft_configured_vpn_server_ip() {
-        let mut ks = KillSwitch::new();
-        let ip: IpAddr = "198.51.100.1".parse().unwrap();
-        ks.set_vpn_server(Some(ip));
-        let rules = ks.build_nft_ruleset(&[]);
-        assert!(rules.contains("ip daddr 198.51.100.1 accept"));
-    }
-
-    #[test]
-    fn test_nft_log_prefix() {
-        let ks = KillSwitch::new();
-        let rules = ks.build_nft_ruleset(&[]);
-        assert!(rules.contains("SHROUD-KS-DROP"));
-    }
-
-    #[test]
-    fn test_nft_input_chain() {
-        let ks = KillSwitch::new();
-        let rules = ks.build_nft_ruleset(&[]);
-        assert!(rules.contains("chain input"));
-        assert!(rules.contains("policy accept"));
-    }
-}
-
-// =========================================================================
 // KillSwitch struct tests (expanded)
 // =========================================================================
 
 #[cfg(test)]
 mod ks_expanded_tests {
     use super::*;
+    use crate::killswitch::paths::ip6tables;
+    use crate::killswitch::rules::DOH_PROVIDERS as DOH_PROVIDER_IPS;
 
     #[test]
     fn test_kill_switch_error_display() {
